Planner integrationg
updated hack-planner and planner to use a file instead of port logic. also updated to work with HACKManager to find best batches
