--- old_scripts/hacking/hack-planner.js
@@ -1,22 +1,0 @@
-import { PortService } from "/scripts/hacking/port-service";
-import { HackPlanner } from "/scripts/hacking/planner";
-
-const FLAGS = [
-    ["help", false],
-    ["port", 4]
-]
-
-/** @param {NS} ns **/
-export async function main(ns) {
-    ns.clearLog();
-
-    const flags = ns.flags(FLAGS);
-    if (flags.help) {
-        ns.tprint("Provide hack/grow/weaken planning services on a netscript port");
-        return;
-    }
-
-    const hackPlanner = new HackPlanner(ns);
-    const service = new PortService(ns, flags.port, hackPlanner);
-    await service.serve();
-}
