+++ ui/batch-view.js
